Fixed Cookie Tests
Updated cookie tests. Now it calculates the cookie expiration date to be some time in the next year. That way it isn't too far into the future. This is what was causing IE and Edge to fail the test. Also, Edge and IE need the domain to not have the initial "." because the driver returns the domain without it, causing the assert to fail.